requirements.txt
python-dotenv>=1.0.0
PyYAML>=6.0.1
pydantic>=2.6.0
langchain-google-genai>=2.0.0
pytest>=7.4.0
